requirements.txt
requests
fuzzywuzzy
python-Levenshtein
nameparser
sqlalchemy
pymysql
lastpass-python
beautifulsoup4
google-cloud-pubsub
google-cloud-secret-manager
google-auth
holidays==0.10.3
colorama
flask
imapclient
html2text
alembic
pandas
numpy
workdays
cryptography
pytest
mock